pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG Chatbot using FastAPI, Cohere, and Qdrant"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "cohere>=4.9.0",
    "qdrant-client>=1.7.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.352",
    "langchain-community>=0.0.352",
    "sentence-transformers>=2.2.2",
    "unstructured>=0.12.5",
    "pypdf>=4.0.0",
    "google-generativeai>=0.8.0",
]
